send happy path working receive echo from echo server working
How do we want the receive api to work? and frankly, the send api as
well...We were headed in the direction of having a run loop that
executes callbacks for messages and errors. We could integrate the send
into that also. Or we could just expect the user to call receive when
they expect something to come in. I think the run loop route was taken
because of bi directional quality of websockets, but I think both can make
sense.